+++ src/main/java/io/proximax/sdk/model/transaction/builder/TransactionBuilderFactory.java
@@ -315,4 +315,37 @@ public class TransactionBuilderFactory {
       initDefaults(builder);
       return builder; 
    }
+
+   /**
+    * create builder for add exchange offer transaction
+    * 
+    * @return the builder
+    */
+   public ExchangeOfferAddTransactionBuilder exchangeAdd() {
+      ExchangeOfferAddTransactionBuilder builder = new ExchangeOfferAddTransactionBuilder();
+      initDefaults(builder);
+      return builder; 
+   }
+
+   /**
+    * create builder for remove exchange offer transaction
+    * 
+    * @return the builder
+    */
+   public ExchangeOfferRemoveTransactionBuilder exchangeRemove() {
+      ExchangeOfferRemoveTransactionBuilder builder = new ExchangeOfferRemoveTransactionBuilder();
+      initDefaults(builder);
+      return builder; 
+   }
+
+   /**
+    * create builder for exchange offer transaction
+    * 
+    * @return the builder
+    */
+   public ExchangeOfferTransactionBuilder exchangeOffer() {
+      ExchangeOfferTransactionBuilder builder = new ExchangeOfferTransactionBuilder();
+      initDefaults(builder);
+      return builder; 
+   }
 }